+++ sdk/device_onboarding_pnp.go
@@ -1872,6 +1872,7 @@ type ResponseDeviceOnboardingPnpGetDeviceByID struct {
 	DayZeroConfigPreview *ResponseDeviceOnboardingPnpGetDeviceByIDDayZeroConfigPreview `json:"dayZeroConfigPreview,omitempty"` // Day Zero Config Preview
 	Version              *float64                                                      `json:"version,omitempty"`              // Version
 	TenantID             string                                                        `json:"tenantId,omitempty"`             // Tenant Id
+	Progress             *ResponseDeviceOnboardingPnpGetDeviceByIDDeviceInfoProgress   `json:"progress,omitempty"`             // Progress
 }
 type ResponseDeviceOnboardingPnpGetDeviceByIDDeviceInfo struct {
 	Source                    string                                                                    `json:"source,omitempty"`                    // Source
@@ -1925,7 +1926,6 @@ type ResponseDeviceOnboardingPnpGetDeviceByIDDeviceInfo struct {
 	PopulateInventory         *bool                                                                     `json:"populateInventory,omitempty"`         // Populate Inventory
 	SiteName                  string                                                                    `json:"siteName,omitempty"`                  // Site Name
 	Name                      string                                                                    `json:"name,omitempty"`                      // Name
-	Progress                  *ResponseDeviceOnboardingPnpGetDeviceByIDDeviceInfoProgress               `json:"progress,omitempty"`                  // Progress
 }
 
 type ResponseDeviceOnboardingPnpGetDeviceByIDDeviceInfoProgress struct {